Optimize the code specification and add comments to the code for reading.
