+++ spring-cloud-alibaba-examples/nacos-example/nacos-discovery-example/nacos-discovery-consumer-example/src/main/java/com/alibaba/cloud/examples/configuration/FeignConfiguration.java
@@ -13,14 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.cloud.examples;
+package com.alibaba.cloud.examples.configuration;
 
+import com.alibaba.cloud.examples.feign.EchoClient;
+import com.alibaba.cloud.examples.feign.EchoClientFallback;
 import org.springframework.context.annotation.Bean;
 
 /**
  * Configuration for Feign.
  *
- * @author HHLJ
+ * @author fangjian0423, MieAh
  */
 public class FeignConfiguration {
