+++ spring-cloud-alibaba-examples/nacos-example/nacos-discovery-example/nacos-discovery-consumer-example/src/main/java/com/alibaba/cloud/examples/configuration/UrlCleaner.java
@@ -13,18 +13,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.alibaba.cloud.examples.configuration;
 
-package com.alibaba.cloud.examples;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+/**
+ * Change the request path containing echo.
+ *
+ * @author fangjian0423, MieAh
+ */
 public class UrlCleaner {
 
-	public static String clean(String url) {
-		System.out.println("enter urlCleaner");
-		if (url.matches(".*/echo/.*")) {
-			System.out.println("change url");
-			url = url.replaceAll("/echo/.*", "/echo/{str}");
-		}
-		return url;
-	}
+    private static final Logger LOGGER = LoggerFactory.getLogger(UrlCleaner.class);
+
+    private static final String URL_CLEAN_ECHO = ".*/echo/.*";
+
+    public static String clean(String url) {
+        LOGGER.info("enter urlCleaner");
+        if (url.matches(URL_CLEAN_ECHO)) {
+            LOGGER.info("change url");
+            url = url.replaceAll("/echo/.*", "/echo/{str}");
+        }
+        return url;
+    }
 
 }
