+++ spring-cloud-alibaba-examples/nacos-example/nacos-discovery-example/nacos-discovery-consumer-example/src/main/java/com/alibaba/cloud/examples/feign/EchoClientFallback.java
@@ -13,26 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.alibaba.cloud.examples.feign;
 
-package com.alibaba.cloud.examples;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
 /**
- * @author xiaojing
+ * When the service is blown, the fallback operation is performed.
+ *
+ * @author fangjian0423, MieAh
  */
-@SpringBootApplication
-@EnableDiscoveryClient
-public class ProviderApplication {
+public class EchoClientFallback implements EchoClient {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ProviderApplication.class, args);
-	}
+    @Override
+    public String echo(@PathVariable("str") String str) {
+        return "echo fallback";
+    }
 
+    @Override
+    public String divide(@RequestParam Integer a, @RequestParam Integer b) {
+        return "divide fallback";
+    }
+
+    @Override
+    public String notFound() {
+        return "notFound fallback";
+    }
 }
