+++ spring-cloud-alibaba-examples/nacos-example/nacos-discovery-example/nacos-discovery-consumer-sclb-example/src/main/java/com/alibaba/cloud/examples/RandomLoadBalancer.java
@@ -1,0 +1,78 @@
+/*
+ * Copyright 2013-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alibaba.cloud.examples;
+
+import org.springframework.beans.factory.ObjectProvider;
+import org.springframework.cloud.client.ServiceInstance;
+import org.springframework.cloud.client.loadbalancer.DefaultResponse;
+import org.springframework.cloud.client.loadbalancer.EmptyResponse;
+import org.springframework.cloud.client.loadbalancer.Response;
+import org.springframework.cloud.loadbalancer.core.NoopServiceInstanceListSupplier;
+import org.springframework.cloud.loadbalancer.core.ReactorServiceInstanceLoadBalancer;
+import org.springframework.cloud.loadbalancer.core.ServiceInstanceListSupplier;
+import reactor.core.publisher.Mono;
+
+import java.util.List;
+import java.util.Random;
+
+/**
+ * Self-defined randomLoadBalancer.
+ *
+ * @author fangjian0423, MieAh
+ */
+public class RandomLoadBalancer implements ReactorServiceInstanceLoadBalancer {
+
+    private ObjectProvider<ServiceInstanceListSupplier> serviceInstanceListSupplierProvider;
+
+    private final String serviceId;
+
+    private final Random random;
+
+    public RandomLoadBalancer(
+            ObjectProvider<ServiceInstanceListSupplier> serviceInstanceListSupplierProvider,
+            String serviceId) {
+        this.serviceInstanceListSupplierProvider = serviceInstanceListSupplierProvider;
+        this.serviceId = serviceId;
+        this.random = new Random();
+    }
+
+    @Override
+    public Mono<Response<ServiceInstance>> choose(
+            org.springframework.cloud.client.loadbalancer.Request request) {
+        ServiceInstanceListSupplier supplier = serviceInstanceListSupplierProvider
+                .getIfAvailable(NoopServiceInstanceListSupplier::new);
+
+        return supplier.get().next().map(this::getInstanceResponse);
+    }
+
+    @Override
+    public Mono<Response<ServiceInstance>> choose() {
+        ServiceInstanceListSupplier supplier = serviceInstanceListSupplierProvider
+                .getIfAvailable(NoopServiceInstanceListSupplier::new);
+        return supplier.get().next().map(this::getInstanceResponse);
+    }
+
+    private Response<ServiceInstance> getInstanceResponse(
+            List<ServiceInstance> instances) {
+        if (instances.isEmpty()) {
+            return new EmptyResponse();
+        }
+        ServiceInstance instance = instances.get(random.nextInt(instances.size()));
+
+        return new DefaultResponse(instance);
+    }
+
+}
