+++ spring-cloud-alibaba-examples/nacos-example/nacos-discovery-example/nacos-discovery-consumer-sclb-example/src/main/java/com/alibaba/cloud/examples/TestController.java
@@ -16,6 +16,7 @@
 package com.alibaba.cloud.examples;
 
 import com.alibaba.cloud.examples.feign.EchoClient;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.client.discovery.DiscoveryClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -26,9 +27,9 @@ import org.springframework.web.client.RestTemplate;
 import javax.annotation.Resource;
 
 /**
- * Example of remote invocation of service fusing and load balancing.
+ * Example of remote invocation of service fusing and custom load balancing.
  *
- * @author xiaojing, fangjian0423, MieAh
+ * @author fangjian0423, MieAh
  */
 @RestController
 public class TestController {
@@ -45,12 +46,26 @@ public class TestController {
     @Resource
     private DiscoveryClient discoveryClient;
 
+    @Value("${spring.cloud.loadbalancer.zone:null}")
+    private String zone;
+
     private static final String SERVICE_PROVIDER_ADDRESS = "http://service-provider";
 
     @GetMapping("/echo-rest/{str}")
     public String rest(@PathVariable String str) {
         return urlCleanedRestTemplate.getForObject(SERVICE_PROVIDER_ADDRESS + "/echo/" + str,
                 String.class);
+    }
+
+    @GetMapping("/zone")
+    public String zone() {
+        return "consumer zone " + zone + "\n" + urlCleanedRestTemplate
+                .getForObject(SERVICE_PROVIDER_ADDRESS + "/zone", String.class);
+    }
+
+    @GetMapping("/echo-feign/{str}")
+    public String feign(@PathVariable String str) {
+        return echoClient.echo(str);
     }
 
     @GetMapping("/index")
@@ -60,12 +75,14 @@ public class TestController {
 
     @GetMapping("/test")
     public String test() {
-        return restTemplate.getForObject(SERVICE_PROVIDER_ADDRESS + "/test", String.class);
+        return restTemplate.getForObject(SERVICE_PROVIDER_ADDRESS + "/test",
+                String.class);
     }
 
     @GetMapping("/sleep")
     public String sleep() {
-        return restTemplate.getForObject(SERVICE_PROVIDER_ADDRESS + "/sleep", String.class);
+        return restTemplate.getForObject(SERVICE_PROVIDER_ADDRESS + "/sleep",
+                String.class);
     }
 
     @GetMapping("/notFound-feign")
@@ -83,11 +100,6 @@ public class TestController {
         return echoClient.divide(a);
     }
 
-    @GetMapping("/echo-feign/{str}")
-    public String feign(@PathVariable String str) {
-        return echoClient.echo(str);
-    }
-
     @GetMapping("/services/{service}")
     public Object client(@PathVariable String service) {
         return discoveryClient.getInstances(service);
@@ -97,5 +109,4 @@ public class TestController {
     public Object services() {
         return discoveryClient.getServices();
     }
-
 }
