+++ spring-cloud-alibaba-examples/nacos-example/nacos-discovery-example/nacos-reactivediscovery-consumer-example/src/main/java/com/alibaba/cloud/examples/WebClientConfiguration.java
@@ -1,0 +1,19 @@
+package com.alibaba.cloud.examples;
+
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.reactive.function.client.WebClient;
+
+/**
+ * Configuration for web client.
+ *
+ * @author fangjian0423, MieAh
+ */
+public class WebClientConfiguration {
+
+    @Bean
+    @LoadBalanced
+    public WebClient.Builder webClient() {
+        return WebClient.builder();
+    }
+}
