+++ spring-cloud-alibaba-examples/nacos-example/nacos-gateway-example/nacos-gateway-provider-example/src/main/java/com/alibaba/cloud/examples/EchoController.java
@@ -1,0 +1,27 @@
+package com.alibaba.cloud.examples;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * 回声控制器
+ *
+ * @author HHLJ
+ * @date 2022/11/29
+ */
+@RestController
+public class EchoController {
+
+    @GetMapping("/echo/{string}")
+    public String echo(@PathVariable String string) {
+        return "hello Nacos Discovery " + string;
+    }
+
+    @GetMapping("/divide")
+    public String divide(@RequestParam Integer a, @RequestParam Integer b) {
+        return String.valueOf(a / b);
+    }
+
+}
